Merge logica

+++ app/Http/Controllers/TeamController.php
@@ -2,6 +2,9 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Bet;
+use App\Models\Game;
+use App\Models\Player;
 use App\Models\Team;
 use Illuminate\Http\Request;
 
@@ -12,7 +15,8 @@ class TeamController extends Controller
      */
     public function index()
     {
-        return Team::all();
+        $teams = Team::with(['league', 'country', 'players'])->get();
+        return response()->json($teams);
     }
 
     /**
@@ -35,6 +39,7 @@ class TeamController extends Controller
         $team->league_id = $request->league_id;
         $team->lost_matches = $request->lost_matches;
         $team->won_matches = $request->won_matches;
+        $team->team_wallpaper = $request->team_wallpaper;
 
         $team->save();
 
@@ -46,7 +51,8 @@ class TeamController extends Controller
      */
     public function show(Team $team)
     {
-        return Team::find($team->id);
+        $team->load(['league', 'country', 'players']);
+        return response()->json($team);
     }
 
     /**
@@ -62,7 +68,7 @@ class TeamController extends Controller
      */
     public function update(Request $request, Team $team)
     {
-        $team -> update($request->all());
+        $team->update($request->all());
         return response()->json($team);
     }
 
@@ -71,6 +77,31 @@ class TeamController extends Controller
      */
     public function destroy(int $team)
     {
-        return Team::destroy($team);
+        $hasBets = Bet::where('team1_id', $team)
+            ->orWhere('team2_id', $team)
+            ->exists();
+
+        $hasGames = Game::where('home_team_id', $team)
+            ->orWhere('away_team_id', $team)
+            ->exists();
+
+        if ($hasBets || $hasGames) {
+            return response()->json([
+                'error' => 'No se puede eliminar el equipo porque forma parte de un partido o de una apuesta existente.'
+            ], 403);
+        }
+
+        Team::destroy($team);
+
+        return response()->json([
+            'message' => 'Equipo eliminado correctamente.'
+        ], 200);
+    }
+
+    public function findTeamsByTeam($teamid)
+    {
+        $players = Player::where('team_id', $teamid)->get();
+
+        return response()->json($players);
     }
 }
